pyproject.toml
[project]
name = "ai-benchmark-leaderboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.17.0",
    "jinja2>=3.1.6",
]
